requirements.txt
openai
autogen
streamlit
python-dotenv
pandas
azure-cognitiveservices-speech
ffmpeg-python

